+++ packages/subtree-updater/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- sync `TreeInsertioEvent`s instead of notes and fill batch events individually
 - sync by TotalEntityIndex instead of block ranges
 - reduce chunk size down from 100000 to 50
 - use merkle index from insertions instead of separate counter